test-data-seeding/requirements.txt
faker==19.6.2
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
